fix: prevent removed plans from being used in workflow commands (#2998)
