+++ renku/command/update.py
@@ -22,7 +22,6 @@ from typing import Optional
 
 from renku.command.command_builder import inject
 from renku.command.command_builder.command import Command
-from renku.command.workflow import execute_workflow
 from renku.core import errors
 from renku.core.errors import ParameterError
 from renku.core.interface.client_dispatcher import IClientDispatcher
@@ -34,6 +33,7 @@ from renku.core.workflow.activity import (
     sort_activities,
 )
 from renku.core.workflow.concrete_execution_graph import ExecutionGraph
+from renku.core.workflow.execute import execute_workflow_graph
 
 
 def update_command():
@@ -81,4 +81,4 @@ def _update(
         return activities, modified_paths
 
     graph = ExecutionGraph([a.plan_with_values for a in activities], virtual_links=True)
-    execute_workflow(dag=graph.workflow_graph, provider=provider, config=config)
+    execute_workflow_graph(dag=graph.workflow_graph, provider=provider, config=config)
