+++ renku/infrastructure/gateway/plan_gateway.py
@@ -17,9 +17,10 @@
 # limitations under the License.
 """Renku plan database gateway implementation."""
 
-from typing import Dict, List, Optional
+from typing import Dict, List, Optional, cast
 
 from renku.command.command_builder.command import inject
+from renku.core import errors
 from renku.core.interface.database_dispatcher import IDatabaseDispatcher
 from renku.core.interface.plan_gateway import IPlanGateway
 from renku.domain_model.workflow.plan import AbstractPlan
@@ -38,9 +39,21 @@ class PlanGateway(IPlanGateway):
         """Get a plan by name."""
         return self.database_dispatcher.current_database["plans-by-name"].get(name)
 
+    def get_by_name_or_id(self, name_or_id: str) -> AbstractPlan:
+        """Get a plan by name or id."""
+        workflow = self.get_by_id(name_or_id) or self.get_by_name(name_or_id)
+
+        if not workflow:
+            raise errors.ParameterError(f'The specified workflow "{name_or_id}" cannot be found.')
+        return workflow
+
     def list_by_name(self, starts_with: str, ends_with: str = None) -> List[str]:
         """Search plans by name."""
-        return self.database_dispatcher.current_database["plans-by-name"].keys(min=starts_with, max=ends_with)
+        return [
+            name
+            for name in self.database_dispatcher.current_database["plans-by-name"].keys(min=starts_with, max=ends_with)
+            if not cast(AbstractPlan, self.get_by_name(name)).deleted
+        ]
 
     def get_newest_plans_by_names(self, include_deleted: bool = False) -> Dict[str, AbstractPlan]:
         """Return a mapping of all plan names to their newest plans."""
