+++ tests/cli/test_workflow.py
@@ -350,6 +350,24 @@ def test_workflow_remove_command(runner, project):
     result = runner.invoke(cli, ["workflow", "remove", "--force", workflow_name])
     assert 0 == result.exit_code, format_result_exception(result)
 
+    result = runner.invoke(cli, ["workflow", "edit", workflow_name, "--name", "new_name"])
+    assert 2 == result.exit_code, format_result_exception(result)
+
+    result = runner.invoke(cli, ["workflow", "execute", workflow_name])
+    assert 2 == result.exit_code, format_result_exception(result)
+
+    result = runner.invoke(cli, ["workflow", "iterate", workflow_name])
+    assert 2 == result.exit_code, format_result_exception(result)
+
+    result = runner.invoke(cli, ["workflow", "compose", "composite", workflow_name])
+    assert 1 == result.exit_code, format_result_exception(result)
+
+    result = runner.invoke(cli, ["workflow", "export", workflow_name])
+    assert 2 == result.exit_code, format_result_exception(result)
+
+    result = runner.invoke(cli, ["workflow", "show", workflow_name])
+    assert 2 == result.exit_code, format_result_exception(result)
+
 
 def test_workflow_export_command(runner, project):
     """Test workflow export with builder."""
@@ -366,7 +384,7 @@ def test_workflow_export_command(runner, project):
     assert len(workflow.outputs) == 1
 
 
-def test_workflow_edit(runner, client, run_shell):
+def test_workflow_edit(runner, client):
     """Test naming of CWL tools and workflows."""
 
     def _get_plan_id(output):
@@ -484,6 +502,22 @@ def test_workflow_edit(runner, client, run_shell):
 
     result = runner.invoke(cli, ["graph", "export", "--format", "json-ld", "--strict"])
     assert 0 == result.exit_code, format_result_exception(result)
+
+
+def test_workflow_edit_no_change(runner, client, run_shell):
+    """Ensure that workflow edit doesn't commit if there's no changes."""
+
+    workflow_name = "my-workflow"
+
+    result = runner.invoke(cli, ["run", "--name", workflow_name, "touch", "data.txt"])
+    assert 0 == result.exit_code, format_result_exception(result)
+
+    before = client.repository.head.commit
+
+    result = runner.invoke(cli, ["workflow", "edit", workflow_name])
+    assert 0 == result.exit_code, format_result_exception(result)
+
+    assert before == client.repository.head.commit
 
 
 def test_workflow_show_outputs_with_directory(runner, client, run):
